delete classes
