+++ applications/app-service/src/main/resources/application.yaml
@@ -63,13 +63,9 @@ jwt:
   secret: AeR9$kL2#pQ7*mN5&vX8!zD4%fG6@hJ3  #${JWT_SECRET} #tu_clave_secreta_muy_larga_y_segura_aqui_1234567890
 
 aws:
-  region: us-east-1
-  endpoint: http://localhost:4566
+  region: us-east-2
   credentials:
-    access-key: test
-    secret-key: test
   sqs:
-    #queue-url: "http://sqs.us-east-1.localhost.localstack.cloud:4566/000000000000/my-queue"
-    queue-url: "http://localhost:4566/000000000000/my-queue"
+    queue-url: "https://sqs.us-east-2.amazonaws.com/703686967236/my-queue"
   sns:
-    topic-arn: arn:aws:sns:us-east-1:000000000000:my-topic
+    topic-arn: arn:aws:sns:us-east-2:703686967236:my-topic
