--- infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/crediya/r2dbc/aws/LambdaInitializer.java
@@ -1,19 +1,0 @@
-package co.com.crediya.r2dbc.aws;
-
-
-import com.amazonaws.services.lambda.runtime.Context;
-import com.amazonaws.services.lambda.runtime.RequestHandler;
-import com.amazonaws.services.lambda.runtime.events.SQSEvent;
-
-public class LambdaInitializer implements RequestHandler<SQSEvent, String> {
-    private final NotificationLambdaHandler handler;
-
-    public LambdaInitializer() {
-        this.handler = new NotificationLambdaHandler();
-    }
-
-    @Override
-    public String handleRequest(SQSEvent event, Context context) {
-        return handler.handleRequest(event, context);
-    }
-}
