+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/crediya/r2dbc/config/AwsConfig.java
@@ -1,60 +1,44 @@
 package co.com.crediya.r2dbc.config;
 
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
-import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
+import software.amazon.awssdk.auth.credentials.DefaultCredentialsProvider;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.sns.SnsClient;
 import software.amazon.awssdk.services.sqs.SqsClient;
 
-import java.net.URI;
-
-@Data
 @Configuration
-@ConfigurationProperties(prefix = "aws")
+@EnableConfigurationProperties(AwsProperties.class)
 public class AwsConfig {
-    private String region;
-    private String endpoint;
-    private AwsCredentials credentials;
-    private SqsConfig sqs;
-    private SnsConfig sns;
+
+    private final AwsProperties awsProperties;
+
+    public AwsConfig(AwsProperties awsProperties) {
+        this.awsProperties = awsProperties;
+    }
 
     @Bean
     public SqsClient sqsClient() {
         return SqsClient.builder()
-                .region(Region.of(region))
-                .credentialsProvider(StaticCredentialsProvider.create(
-                        AwsBasicCredentials.create(credentials.getAccessKey(), credentials.getSecretKey())))
-                .endpointOverride(URI.create(endpoint))
+                .region(Region.of(awsProperties.getRegion()))
+                .credentialsProvider(DefaultCredentialsProvider.create())
                 .build();
     }
 
     @Bean
     public SnsClient snsClient() {
         return SnsClient.builder()
-                .region(Region.of(region))
-                .credentialsProvider(StaticCredentialsProvider.create(
-                        AwsBasicCredentials.create(credentials.getAccessKey(), credentials.getSecretKey())))
-                .endpointOverride(URI.create(endpoint))
+                .region(Region.of(awsProperties.getRegion()))
+                .credentialsProvider(DefaultCredentialsProvider.create())
                 .build();
     }
 
-    @Data
-    public static class AwsCredentials {
-        private String accessKey;
-        private String secretKey;
+    public AwsProperties.SqsConfig getSqs() {
+        return awsProperties.getSqs();
     }
 
-    @Data
-    public static class SqsConfig {
-        private String queueUrl;
-    }
-
-    @Data
-    public static class SnsConfig {
-        private String topicArn;
+    public AwsProperties.SnsConfig getSns() {
+        return awsProperties.getSns();
     }
 }
